requirements.txt
# =============================================================================
# APS_Project (Auto Photo Studio) - Python Dependencies
# =============================================================================
# Install via:  pip install -r requirements.txt
# NOTE: ComfyUI itself is NOT pip-installed; it is cloned separately and
#       invoked headlessly.  The Impact-Pack node is installed inside ComfyUI's
#       custom_nodes directory.  This file covers the *pipeline orchestrator*
#       only.
# =============================================================================

# --- Image Quality Assessment (The Sorter) ---
# Rank-IQA reference implementation; fallback to PIQE (OpenCV contrib).
# pyiqa wraps dozens of NR-IQA metrics behind a single API.
pyiqa>=0.1.10

# --- Object / Face Detection (The Mapper) ---
# Ultralytics YOLOv8 for face bounding-box detection.
ultralytics>=8.1.0

# --- Semantic Segmentation (The Mapper) ---
# BiSeNet-V2 skin/hair/background parsing.
# We load via torch + a lightweight model zoo helper.
torch>=2.1.0
torchvision>=0.16.0

# --- General image I/O ---
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# --- RAW file decoding (Sony ARW / A7M5) ---
rawpy>=0.19.0

# --- ComfyUI API communication (The Forge) ---
websocket-client>=1.6.0
requests>=2.31.0
aiohttp>=3.9.0

# --- Pipeline utilities ---
tqdm>=4.66.0
pyyaml>=6.0.1

# --- Darkroom API server ---
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
